Move getExcelFile to write util file
